Hidden setting to override the text editor.
Note that people who run gtimelog straight from Git will need to
manually run ``make`` to recompile the gsettings schema, or gtimelog
will segfault when you press Ctrl+E.

Also note that using `gsettings` is a pain when the schema is not
installed globally, but you can use

    dconf write /org/gtimelog/editor '"gvim %s"'

and then tweak it with dconf-editor.

Error handling is missing: if you set editor to gvim and gvim is not
installed, you'll see a Python traceback on the console.

Application.prepare_args() is a separate static method so I could write
unit tests for it, but I haven't done so.

Closes #113.